requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
astrapy==2.1.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
dataclasses-json==0.6.7
datasets==4.1.1
ddgs==9.6.0
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
executing==2.2.1
fastapi==0.118.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
groq==0.31.1
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
instructor==1.11.3
ipykernel==6.30.1
ipython==9.5.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.27
langchain-astradb==0.6.1
langchain-community==0.3.30
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-groq==0.3.8
langchain-mcp-adapters==0.1.11
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.28
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
multidict==6.6.4
multiprocess==0.70.16
mypy-extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
openai==1.109.1
orjson==3.11.3
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
primp==0.15.0
prompt-toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pydeck==0.9.1
pygments==2.19.2
pymongo==4.15.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
ragas==0.3.6
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-network==0.33.3
scipy==1.16.2
selenium==4.35.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.8
sqlalchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
streamlit==1.49.1
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.19.2
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
